-- Changed the numerical data type used by the memory manager (hunk/zone) from `signed int` to `size_t` to remove the 2GB hard limit in the code.     *Note: the `%zu` format specifier wasn't used because of MSVCRT shenanigans, but it should be once a fix has been found.* -- Added `--large-address-aware` linker flag to the Makefile when using MinGW and 32-bit to slightly increase the available virtual address space for 32-bit apps on 64-bit Windows. -- Added Makefile variable to set the language standard. -- Language standard has been bumped from C89 to C17 (not applicable to QVMs, since LCC is C89-only). -- Added `.gitignore`.

+++ Engine/renderercommon/tr_public.h
@@ -119,11 +119,11 @@ typedef struct {
 	// stack based memory allocation for per-level things that
 	// won't be freed
 #ifdef HUNK_DEBUG
-	void	*(*Hunk_AllocDebug)( int size, ha_pref pref, char *label, char *file, int line );
+	void	*(*Hunk_AllocDebug)( size_t size, ha_pref pref, char *label, char *file, int line );
 #else
-	void	*(*Hunk_Alloc)( int size, ha_pref pref );
+	void	*(*Hunk_Alloc)( size_t size, ha_pref pref );
 #endif
-	void	*(*Hunk_AllocateTempMemory)( int size );
+	void	*(*Hunk_AllocateTempMemory)( size_t size );
 	void	(*Hunk_FreeTempMemory)( void *block );
 
 	// dynamic memory allocator for things that need to be freed
